Refactor subscriptions to use async and reuse common code
